requirements.txt
selenium==4.16.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
requests==2.31.0
tqdm==4.66.1
lxml==4.9.3
psycopg2-binary==2.9.9
schedule==1.2.1
python-dotenv==1.0.0
argparse==1.4.0
aiohttp==3.9.1
prometheus-client==0.19.0
structlog==24.1.0
tenacity==8.2.3
redis==5.0.1
jsonschema==4.21.1
pandas==2.2.0